# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/crm/accounts/{tenantId}": {
      "get": {
        "operationId": "CrmController_getAccounts",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/crm/contacts/{tenantId}": {
      "get": {
        "operationId": "CrmController_getContacts",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/crm/leads/{tenantId}": {
      "get": {
        "operationId": "CrmController_getLeads",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/crm/notes/{contactId}": {
      "get": {
        "operationId": "CrmController_getNotes",
        "parameters": [
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/purchasing": {
      "get": {
        "operationId": "PurchasingController_getAll",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PurchasingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/purchasing/{id}/receive": {
      "put": {
        "operationId": "PurchasingController_receive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "purchasing"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "HVAC API",
    "description": "Generated API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePoDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "skuId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "skuId",
          "quantity"
        ]
      }
    }
  }
}